Added Mastodon verification for my other accounts
Due to my work on muenchen.social and troet.cafe I needed to add these lines to verify my identity for them as well. Also I removed the word "Mastodon" as it is not necessary, I can't believe I worked with this stupid "display:none" when I could've just removed the word. "rel="me"" is what is important, but whatever.